requirements.txt
django>=1.7
requests
requests-oauthlib
googler
django-jsonfield
python-dateutil
dnspython3
netaddr